Baseline functionality
This introduces a basic baseline functionality similar to PHPStan. Except that it
uses the `baselineErrors` key in the configuration.

You can use --generate-baseline to generate the baseline and include that.

There is a `reportUnmatchedIgnoredErrors` option to control what should happen
when the baseline error does not match.

Note that `reportUnmatchedIgnoredErrors` does not yet work for `ignoreErrors`.

Closes #18
Fixes #9