Do not convert `$context` access to variables
When referencing a variable, Twig fetches it from the `$context` array like so:

```twig
{{ name }}
```

Becomes this in PHP:
```php
echo $context['name'] ?? null;
```

Before, we would convert all these array based fetches to variables:
```php
echo $name;
```

The purpose of this was to cleanup the errors.

It would say:
```
Variable $name might not exist.
```

Instead of
```
Offset 'name' does not exist in array{other: string, things: 'here'}
```

But doing all this means that we further drift from what Twig normally does.

Maybe it's not so bad that the error talks about offset and then shows the context.
Next to that, we can try to improve those error messages later.