Collect context for include calls in Twig (#60)
Whenever you include a template inside a Twig template, we can collect the
context that is passed to the template.

We can then inject this into the compiled template so that we have type information there too.

This is still limited, as this currently only collects the variables defined in the template
that includes it.

We should think about how to get the context that is passed from the controller. Maybe this will
require another PHPStan run.