Improve performance (#67)
* Remove comment

* Remove nullability from twigstan PHPStan configuration

* Ignore errors in JSON when only collecting

When we're collecting, we don't do anything with the errors. By ignoring them, we safe time encoding
and decoding it later.

* Prettify dumped JSON when in debug mode

* Use block mode for Neon configuration

* Use real path to phpstan.neon

* Keep track of PHPStan files in tempDirectory

This includes the configuration file, PHPStan's tmp directory and result caches.

* Change render point emoji

The old one doesn't render nicely in my terminal Ghostty.

* Always return with non-zero exit code

* Do not discard notFileSpecificErrors when collecting

* Keep track of PHPStan exit code

* Allow specifying files in twig and php paths

* Merge union types when possible

Otherwise this could potentially create very complex union PHPDocs that PHPStan fails to handle.

* Define paths to analyze in PHPStan configuration instead of as argument

Before, we would pass a list with all files. On large projects this would exceed the maximum limit
of arguments.

Next to that, the result cache would not be created because of that.

To solve this, we now overwrite the paths that PHPStan should analyze.

Fixes #8

* Remove redundant temp directory config in tests

* Ignore missingType.iterableValue on main/block functions