Do not simplify templates
The more we drift from the actual compiled template, the harder it is to maintain.

By doing this, the errors when calling a macro are gone.

This is not full support for macros yet, but at least no false positive errors.

Next to that, this changes the way TwigStan interacts with PHPStan.

It now first collects contexts from PHP files. The templates are then compiled with
this context initially set. Globals are no longer compiled but loaded dynamically via
the GetGlobalsReturnType extension.

A new configuration option `onlyAnalyzeTemplatesWithContext` has been introduced for people
that want to exclude errors for templates that have no render point.

When analysis is done by specifying the files via CLI, TwigStan now detects imported macros as a
dependency for the template and makes sure that template is also analyzed.