Support included templates
This introduces a new mechanism that allows TwigStan to re-run full analysis for certain templates.
In this case, it detects when a template is included in a template. It will take the context at that
point and analyzes the template with that context.

Obviously this slows down analysis, but gives less false positives.

Fixes #54